Main and streamlit
